+++ lib/src/widgets/animated_text_form_field.dart
@@ -2,8 +2,8 @@ import 'dart:math';
 
 import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_login/flutter_login.dart';
-import 'package:flutter_login/src/widgets/term_of_service_checkbox.dart';
+import 'package:flutter_login_MG/flutter_login.dart';
+import 'package:flutter_login_MG/src/widgets/term_of_service_checkbox.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 import 'package:intl_phone_number_input/intl_phone_number_input.dart';
 import 'package:phone_numbers_parser/phone_numbers_parser.dart' as pnp;